chore: move service into its own package and separate it from the kubernetes api types
