+++ pkg/service/service.go
@@ -1,19 +1,20 @@
-package v1alpha1
+package service
 
 import (
 	"context"
 	"errors"
 
+	"github.com/openmfp/account-operator/api/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 type Service interface {
-	GetFirstLevelAccountForAccount(ctx context.Context, accountKey client.ObjectKey) (*Account, error)
-	GetFirstLevelAccountForNamespace(ctx context.Context, namespace string) (*Account, error)
+	GetFirstLevelAccountForAccount(ctx context.Context, accountKey client.ObjectKey) (*v1alpha1.Account, error)
+	GetFirstLevelAccountForNamespace(ctx context.Context, namespace string) (*v1alpha1.Account, error)
 
-	GetAccount(ctx context.Context, accountKey client.ObjectKey) (*Account, error)
-	GetAccountForNamespace(ctx context.Context, namespace string) (*Account, error)
+	GetAccount(ctx context.Context, accountKey client.ObjectKey) (*v1alpha1.Account, error)
+	GetAccountForNamespace(ctx context.Context, namespace string) (*v1alpha1.Account, error)
 }
 
 var _ Service = (*service)(nil)
@@ -41,12 +42,12 @@ func (s *service) getAccountOwnerAndNamespaceForNamespace(ctx context.Context, n
 		return "", "", errors.New("namespace does not have a label and therefore no connected account")
 	}
 
-	accountNamespace, ok := ns.Labels[NamespaceAccountOwnerNamespaceLabel]
+	accountNamespace, ok := ns.Labels[v1alpha1.NamespaceAccountOwnerNamespaceLabel]
 	if !ok || accountNamespace == "" {
 		return "", "", errors.New("namespace does not have an account-owner-namespace label and therefore no connected account")
 	}
 
-	accountName, ok := ns.Labels[NamespaceAccountOwnerLabel]
+	accountName, ok := ns.Labels[v1alpha1.NamespaceAccountOwnerLabel]
 	if !ok || accountName == "" {
 		return "", "", errors.New("namespace does not have an account-owner label and therefore no connected account")
 	}
@@ -54,11 +55,11 @@ func (s *service) getAccountOwnerAndNamespaceForNamespace(ctx context.Context, n
 	return accountName, accountNamespace, nil
 }
 
-func (s *service) GetFirstLevelAccountForAccount(ctx context.Context, accountKey client.ObjectKey) (*Account, error) {
+func (s *service) GetFirstLevelAccountForAccount(ctx context.Context, accountKey client.ObjectKey) (*v1alpha1.Account, error) {
 	return s.GetFirstLevelAccountForNamespace(ctx, accountKey.Namespace)
 }
 
-func (s *service) GetFirstLevelAccountForNamespace(ctx context.Context, namespace string) (*Account, error) {
+func (s *service) GetFirstLevelAccountForNamespace(ctx context.Context, namespace string) (*v1alpha1.Account, error) {
 
 	accountName, accountNamespace, err := s.getAccountOwnerAndNamespaceForNamespace(ctx, namespace)
 	if err != nil {
@@ -69,24 +70,24 @@ func (s *service) GetFirstLevelAccountForNamespace(ctx context.Context, namespac
 		return s.GetFirstLevelAccountForNamespace(ctx, accountNamespace)
 	}
 
-	var account Account
+	var account v1alpha1.Account
 	err = s.client.Get(ctx, client.ObjectKey{Name: accountName, Namespace: accountNamespace}, &account)
 	return &account, err
 }
 
-func (s *service) GetAccount(ctx context.Context, accountKey client.ObjectKey) (*Account, error) {
-	var account Account
+func (s *service) GetAccount(ctx context.Context, accountKey client.ObjectKey) (*v1alpha1.Account, error) {
+	var account v1alpha1.Account
 	err := s.client.Get(ctx, accountKey, &account)
 	return &account, err
 }
 
-func (s *service) GetAccountForNamespace(ctx context.Context, namespace string) (*Account, error) {
+func (s *service) GetAccountForNamespace(ctx context.Context, namespace string) (*v1alpha1.Account, error) {
 	accountName, accountNamespace, err := s.getAccountOwnerAndNamespaceForNamespace(ctx, namespace)
 	if err != nil {
 		return nil, err
 	}
 
-	var account Account
+	var account v1alpha1.Account
 	err = s.client.Get(ctx, client.ObjectKey{Name: accountName, Namespace: accountNamespace}, &account)
 	return &account, err
 }
